Lastly conform to the specification by using size_t instead of int for the sizes.

+++ include/stdlib.h
@@ -11,7 +11,7 @@ int abs(int n);
 
 void free(void *p);
 void *malloc(size_t sz);
-void *calloc(int elements, int elementSize);
+void *calloc(size_t elements, size_t elementSize);
 void *realloc(void *p, size_t sz);
 
 int rand(void);
